refactor: move python bindings for values into separated files
