+++ src/ffi/py/mod.rs
@@ -3,7 +3,6 @@ pub(crate) mod cache_progress;
 #[cfg(feature = "ailoy-model-cli")]
 pub(crate) mod cli;
 pub(crate) mod string_enum;
-pub(crate) mod value;
 
 use pyo3::prelude::*;
